requirements.txt
# Backend Dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.1.0
openai==0.28
typing-extensions>=4.5.0
PyPDF2==3.0.0
tiktoken==0.5.2  # Added tiktoken

# Frontend Dependencies
streamlit==1.40.2
streamlit-authenticator==0.3.1
streamlit-option-menu==0.3.12
pandas==2.2.0
plotly==5.18.0
requests==2.31.0
python-jose==3.3.0
bcrypt==4.1.2
PyYAML==6.0.1
extra-streamlit-components==0.1.60
fpdf2==2.5.1

# RAG System Dependencies
numpy==1.26.0
jsonschema==4.19.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.33.0
chromadb==0.4.22  # Added ChromaDB
#hnswlib>=0.7.0    # Required for ChromaDB
sqlalchemy>=2.0.0 # Required for ChromaDB persistence

# Web Scraping Dependencies
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.25.1

# Transcribe
python-multipart==0.0.19
streamlit-audiorec==0.1.3
